Reorganize the directory structure
It seems that a lot of the problems I've had with runtime paths
in this package stem from the fact that this package "adds" files
to the tzinfo collection and wasn't putting them in a
subdirectory that had a distinct path from the collection root
from the files that are provided by the tzinfo package itself.

So this PR moves the zoneinfo files into the collection path:
   tzinfo/tzdata/zoneinfo

It also defines a new key in the info.rkt file containing this
path, for use by the tzinfo package. It also defines the older
variable, too, for use by older versions of tzinfo.

In addition, the build.rkt script is updated, and we're now using
the 2019c version of tzdb.