Copy zoneinfo files into shared directory on package install
This will allow the tzinfo package to define a runtime path
to the files, allowing them to be used in standalone applications.
The current way that tzdata communicates the location of these
files is via `find-relevant-directories`, but that doesn't
work in the context of a standalone executable.

See the following racket-users mailing list discussion for
details:
https://groups.google.com/g/racket-users/c/_2lHI5SkqJo/m/XubQz-waBgAJ