Build "fat" tzfiles
At some point, the default tzdb build started to produce "slim"
files, which (as the name suggests) are smaller, but they do not
include future DST transitions. They only produce transitions up
to the last point in time when the rules governing DST transitions
changed. For example, in America/New_York, this was March 2007.

Since the tzinfo library only uses transitions that are explicit
in tzfiles, it produces incorrect results for slim files.

I intend to change tzinfo so that it can use slim files in the
future, but a quick workaround is to go back to using fat files.

See: https://github.com/97jaz/tzinfo/issues/18